Add IO!
Well, sort of.  Currently the only primitives are exitFailure and
testAll, just enough to make test/a.dk into a set of unit tests.

Currently, the IO monad is exposed through dedicated Ir constructs
(Bind, Return, PrimIO).  This is an alternative to the RealWorld
passing of GHC.  I made this choice because it seems more similar
to effect systems in spirit, but that's an extremely vague notion
which may turn out wrong/useless with time.