First draft of type inference
At this point, type inference compiles but loops forever on almost
everything, on crashes on almost everything else.  The next few
commits will fill in the missing pieces, which are (roughly):

1. The results of function type inference are not cached.
2. There is no fixed point iteration to resolve recursive function types.
3. Distinct overloaded functors (function types) fail to unify.
4. There is no higher order function type generalization step.

The reason this is worthy of its own commit is that the code is
a wonderfully exact analog to the interpreter.  In fact, I generated
most of Infer.hs by copying Interp.hs, applying slight modifications to
the type signatures, and fixing the resulting compile errors.  The main
conceptual changes appear in the handling of IO, Cons, and Case.