Add function type caching and fixpoint iteration to type inference
Function type unification needs to be fixed next.  Currently it always
succeeds (the first draft commit message was wrong) and returns one of
the arguments arbitrarily rather than performing true unification.

The current fixpoint iteration is suboptimal, since it unconditionally
does one extra iteration in order to detect the fixed point.  It should
be possible to terminate after only one iteration in the case of
recursive functions (and more generally for recursive functions with
precise signatures).