Split Type into Type and TypeSet
This moves us closer to the correct handling of function types during
type inference, since it clarifies which types are concrete and which
are not.  Specifically, Type represents a single concrete type, and
therefore has no TyVar constructor, and TypeSet represents a set of
types for purposes of overload specification.

Other miscellaneous improvements in this commit include:

1. Factor Callstack out of ExecMonad and InferMonad into a separate file.
2. Thread type information through the exec monad to make it available
   to the interpreter.
3. Add a MonadInterrupt class encapsulating asynchronous exception
   handling in the Exec and Infer monads.
4. Thread function type lookup and inference through the various type
   unification functions in Type.hs.
5. Add an ErrorT component to the Infer monad to support rollback of
   errors for SFINAE purposes.
6. Rename unifyS to intersect/union to reflect the set theoretic
   interpretation.
7. Generalize TyClosure to allow a finite union of different closure
   types so that the "intersect" function is appropriately closed.
8. Add a MaybeT monad transformer to handle simple success/failure cases.

The main flaw in function type unification is that information flows in
only one direction when unifying lists of types or subtypes.  Therefore,
function type sets like (a -> b) -> [a] -> [b] will fail to unify.  This
will be fixed in a subsequent commit.