Make unification order adaptive
The order in which subtypes are unified must be adaptive in the presence
of function type sets.  For example, when unifying (a -> Int, a) with
(Negate, Int), the value of "a" must be determined from the second part
of the tuple before the function part can be checked.  To handle this,
unify' produces a list of indeterminate subcomputations as it runs, and
unify iterates on this until a fixed point is reached.

I also fixed a bug in fixpoint iteration in cache, which means that type
inference now succeeds for basic.duck!  The next step is eliminating
typeof in the interpreter.