Remove \ from anonymous function syntax
The syntax for anonymous functions is now the same as the syntax for
function types, which is an important step towards unifying type and
expression syntax completely.  The resulting grammar is not technically
LALR(1), so we parse anonymous functions as "exp -> exp" and then
reparse the first expression into a pattern.

Since each "->" corresponds to exactly one argument, we can drop
parentheses in cases like \(h:t) -> e.  This convenience more than
makes up for the extra arrows required for multiple argument lambdas.

Finally, we reuse the now unnecessary backslash as a slightly more
powerful version of Haskell's $ operator (it works for tuples as well).

Minor changes in this commit include

1.  Add list patterns, e.g., [a,b,c].
2.  Rename TypeCheck to Spec and implement in Ir and Lir.  Types are
    still separate syntax and don't participate in variable scoping,
    so more work will be required.
3.  Update Lir pretty printing (no more "over").
4.  Set -Werror in Makefile.
5.  Do the nonterminal duplication necessary to avoid all shift-reduce
    conflicts in the grammar.
6.  Remove commas from precedence declaration syntax.