Merge adjacent '->'s into a single Ast.Lambda
The merging is dependent on parentheses: x -> x -> x + x becomes

    Lambda [x,x] (x+x)

and x -> (x -> x + x) becomes

    Lambda [x] (Lambda [x] (x+x))

This will have semantic effect once duplicate variables in pattern
matches are fixed to produce unification constraints.  The meanings
will then become

    x -> x -> x + x   -- pattern match fails if the arguments differ
    x -> (x -> x + x) -- first argument is ignored

The motivation for this is the unification of type and expression
syntax; t -> t -> t takes two arguments of the same type.  This
behavior is a bit magical, so it might still be changed if a better
solution is found.