Merge type and expression sections of the grammar
The types are immediately reparsed from expressions into type objects,
so the merging of types and expressions isn't complete.  Closer, though.

Since function types like (t -> t) -> t are now parsed into expressions
first, and the left-hand side of -> expressions are patterns, we now
have a lambda constructor for patterns (PatLambda).  This is currently
unimplemented (right now it serves only as an intermediate
representation for types in the grammar).  It will definitely become
implemented for type patterns once the merging is complete, and we may
or may not want to implement the analogous behavior for value patterns
(it amounts to a backwards way of writing function application, and is
pure syntactic sugar).

The addition of the unimplemented PatLambda's caused all the pattern
matching conversion code in Ir.hs to sometimes through exceptions,
which required significant restructuring.  This might have pushed me
over the edge in favor of believing that effect types are a better
default than monads.

The precedence of ',' is now unconditionally higher than that of '->'.
Previously this was true for expressions not but for types.