Disallow duplicate variable definitions in patterns
I spent several days trying to merge the unification performed on
typesets for function type specifications with expression pattern
matching, and failed.  The main issue is the strange recursive
nature of type unification; unification with a typeset like
(a -> b) -> (b -> a) -> a -> a can potentially loop forever, which
is not something we want in expression pattern matching.

As long as these are separate, there's no pressing need to implement
implicit equality constraints for duplicate variables in a pattern,
so this commit disallows them.