Revap function type representation
Function types are now represented by a single Type (TypeSet)
constructor TyFun (TsFun) wrapping around a TypeFun type.  TypeFun
represents an intersection of any number of arrows and closures.

This was motivated by the discovery of several huge bugs in the
handling of function types:

1. "union" computed the same thing as "intersect" for closure types.
2. Intersections of arrow and closure types were computed too
   conservatively, so no principle types for us.  In particular,
   "intersect (a -> b) f" failed if "f a" was invalid.  This is inexact
   If there is a proper subtype c of a s.t. "f c" is valid.

The logic for intersecting, unioning, and unifying function types is
now split out into separate functions.  It still has various bugs, in
particular due to the lack of rollback ability in "apply", but these
can be fixed later.