Annotate each datatype type argument with covariant/invariant
Each type argument to a type constructor is treated as either covariant
or invariant.  A covariant argument occurs as concrete data, while an
invariant type appears as an argument to a function (or similar).  For
example, in

    data T x y of A x y (x -> Int)

y is covariant but x is invariant.  In terms of subtyping, we have

    T Int Void <= T Int Int   -- since Void <= Int
    T Void Int !<= T Int Void -- fails; the first argument is invariant

Previously, all type constructor arguments were implicitly assumed
covariant, which is clearly unsound.