Implement Type (x,y) <-> (Type x, Type y) equivalence
This adds the bare minimum amount of type constructor coercion to allow
tuples to be passed to functions that expect tuple types.  Because the
current mechanism assumes perfect symmetry, it will have to be massaged
a bit in order to make lists work.

At some point in the future, we'll probably want to clean up and
generalize this support, including not hard coding the set of type
constructor coercions.