Switch to raw memory layout
Duck values are now stored in a (simplistic) native format, rather
than as a Haskell algebraic datatype.  This is a first step towards
having a real backend.

The original plan was to use the Boehm Garbage Collector for these
values, but unfortunately Boehm appears to conflict with the Haskell
runtime system.  Thus, for now, we run without a garbage collector
and leak all memory.

The current memory layout is extremely simple: no unboxing of zero
arity constructors, no zero size datatypes, etc.