Pull type transform annotations into arrow types
This allows Trans types to be correctly inferred.  For now (and possibly ever),
Delayed is strictly incompatible with NoTrans, so delaying functions cannot be
passed as arguments where non-delaying functions are expected.

This only affects one test-case, fold (||), which has a strange-looking fix
(and a new working overload test case in delay.duck that would also fix it).
This suggests that the syntax and overload methodology of Delayed should be
reconsidered, but the current semantics may be reasonable.