Change transform syntax to "delay a -> ..."
This expands the parser to allow applications in patterns and types, which are
converted to proper Trans representation in Ir and Lir respectively (the
assymetry here is unfortunate, but patterns must be converted during Ir, and
types are too messy to do before Lir).

This also makes the resulting "Delay a" type abstract, though force is still
achieved by applying it to (), which may be better handled with a primitive.
This also allows ValDelay to be distinguished from ValClosure before runtime,
eliminating an extra tag layer.

Also added framework for allowing "delay a" to implicity provide an overload
for "Delay a" as well, but this turned out to be a bad idea in practice, as
"if :: Bool -> Delay a -> Delay a -> ?" is intuitively ambiguous, so this is
commented out and should perhaps be revisited or removed.