upgrade to alex 3.0
Should be able to add alex version specification to .cabal but I don't know how.
Use ByteString to simplify encoding issues.
This breaks utf-8 encoding of tokens for the moment (and possibly DOS newlines, but this is a known bytestring issue).
Does not eliminate the warning that prompted me to try this in the first place, however.